requirements.txt
Django==4.1.3
python-dotenv==0.21.0
django-phonenumber-field[phonenumberslite]
requests==2.28.1

